added longpolling support for pages that display locales.
discussions are now enabled for topics, the third tier in the hierarchy. however, the longpolling support is somehow flawed, and for the life of me I can't figure out why. it works perfectly for 5 comments, and then stops working. The comments aren't recorded yet as there's nothing in the API that lets me do that right now, but they (the first 5) are working perfectly. I suspect it's an issue on the server side, but I havent been able to pinpoint it.